+++ src/main/java/step5/UsingNameRacingCarGame.java
@@ -1,10 +1,10 @@
-package step4;
+package step5;
 
-import step4.domain.UsingNameCarRacingInformation;
-import step4.domain.UsingRacingCarStage;
-import step4.strategy.move.MoveStrategy;
-import step4.view.InputView;
-import step4.view.UsingNameCarPrintView;
+import step5.domain.UsingNameCarRacingInformation;
+import step5.domain.UsingRacingCarStage;
+import step5.strategy.move.MoveStrategy;
+import step5.view.InputView;
+import step5.view.UsingNameCarPrintView;
 
 public class UsingNameRacingCarGame extends AbstractRacingCarGame<UsingNameCarRacingInformation> {
