+++ src/main/java/step5/domain/UsingNameCarRacingInformation.java
@@ -1,4 +1,4 @@
-package step4.domain;
+package step5.domain;
 
 import java.util.*;
 import java.util.stream.Collectors;
