+++ src/main/java/step5/domain/UsingNameRacingCar.java
@@ -1,6 +1,6 @@
-package step4.domain;
+package step5.domain;
 
-import step4.strategy.move.MoveStrategy;
+import step5.strategy.move.MoveStrategy;
 
 public class UsingNameRacingCar {
