+++ src/main/java/step5/strategy/move/RandomMoveStrategy.java
@@ -1,4 +1,4 @@
-package step4.strategy.move;
+package step5.strategy.move;
 
 import java.util.concurrent.ThreadLocalRandom;
